fixed movement for real this time and it should work for you too
