+++ 169Capstone/Assets/Prefab/Enemies/Slime Enemy.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a76be535b1a8ef643a209f253f070d83
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
