+++ 169Capstone/Assets/ScriptableObjects/Enemies/Logic Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 315106daa43594448847a69acb3a94bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
