+++ 169Capstone/Assets/ScriptableObjects/Enemies/Logic Objects/Robert Logic.asset.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 0a5083fe0e83f430fa0432b92e001d6a
-PrefabImporter:
+guid: c7d1158479dcde343a2b97425ae9c5fc
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
