+++ 169Capstone/Assets/ScriptableObjects/Enemies/Logic Objects/Slime Logic.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4e3d51ac7bd3ff43af6c95c7ee4e206
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
