+++ 169Capstone/Assets/ScriptableObjects/Enemies/Robert.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8488a872beb2af147a09495933bc5289
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
