+++ 169Capstone/Assets/Scripts/Entity/Enemy/Enemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dc9df3ef0c296427596e6808f1ea59b2
+guid: a83fe560de6540047bd32a5885d70f14
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
