+++ 169Capstone/Assets/Scripts/Entity/Enemy/EnemyLogic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b43ba18994717245b332e11cd63188a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
