+++ 169Capstone/Assets/Scripts/Entity/Enemy/Robert.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb1e4dce67155044b875c84cc3ac34e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
