+++ 169Capstone/Assets/Scripts/Entity/Health.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c68a0f0c842e406419e4b29c6d83ed5c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
