+++ 169Capstone/Assets/Scripts/Entity/Player/Movement/Movement.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 public class Movement : MonoBehaviour
 {
@@ -8,38 +9,88 @@ public class Movement : MonoBehaviour
     public Transform model;
     public CharacterController player;
     public float smoothing = 0.1f;
-
     private float smoothingVelocity;
-    // Start is called before the first frame update
+    
+    private float horizontalMove;
+    private float verticalMove;
+
+    public Animator animator;
+    
     void Start()
     {
-        
+        InputActionAsset controls = gameObject.GetComponent<PlayerInput>().actions;
+
+        // Add STOPPING moving when you're no longer holding the button
+        controls.FindAction("MoveHorizontal").canceled += x => OnMoveHorizontalCanceled();
+        controls.FindAction("MoveVertical").canceled += x => OnMoveVerticalCanceled();
     }
 
-    // Update is called once per frame
     void Update()
     {
-        float vertical = Input.GetAxisRaw("Horizontal");
-        float horizontal = Input.GetAxisRaw("Vertical");
-        Vector3 direction = new Vector3(-horizontal, 0, vertical).normalized;
+        if(!InputManager.instance.CanAcceptGameplayInput()){
+            return;
+        }
+        HandleMovement();
+    }
+
+    public void OnMoveHorizontal(InputValue input)
+    {
+        if(!InputManager.instance.CanAcceptGameplayInput()){
+            return;
+        }
+        verticalMove = input.Get<float>();
+        animator.SetBool("IsRunning", true);
+    }
+
+    public void OnMoveHorizontalCanceled()
+    {
+        verticalMove = 0;
+        CheckForIdle();
+    }
+
+    public void OnMoveVertical(InputValue input)
+    {
+        if(!InputManager.instance.CanAcceptGameplayInput()){
+            return;
+        }
+        horizontalMove = input.Get<float>();
+        animator.SetBool("IsRunning", true);
+    }
+
+    public void OnMoveVerticalCanceled()
+    {
+        horizontalMove = 0;
+        CheckForIdle();
+    }
+
+    private void CheckForIdle()
+    {
+        if(verticalMove == 0 && horizontalMove == 0){
+            animator.SetBool("IsRunning", false);
+        }
+    }
+
+    private void HandleMovement()
+    {
+        Vector3 direction = new Vector3(-horizontalMove, 0, verticalMove).normalized;
 
         if(direction.magnitude >= 0.1f)
         {
             float rad = 45 * Mathf.Deg2Rad;
-            if(horizontal > 0)
+            if(horizontalMove > 0)
             {
                 direction.z -= rad;
             }
-            else if(horizontal < 0)
+            else if(horizontalMove < 0)
             {
                 direction.z += rad;
             }
 
-            if(vertical > 0)
+            if(verticalMove > 0)
             {
                 direction.x -= rad;
             }
-            else if(vertical < 0)
+            else if(verticalMove < 0)
             {
                 direction.x += rad;
             }
